requirements.txt
Django==1.5
Pillow==2.4.0
South==0.7.4
argparse==1.2.1
django-wsgiserver==0.6.9
psycopg2==2.5.1
simplejson==3.4.0
wsgiref==0.1.2
django-social-auth==0.7.28
django-categories==1.2.2

